feat : 검색페이지 -> 상세페이지 연결
