+++ src/pages/movies/[movieId].js
@@ -1,0 +1,11 @@
+import { useRouter } from "next/router";
+
+export default function Detail() {
+  const router = useRouter();
+
+  return (
+    <div>
+      <h4>{router.query.title}</h4>
+    </div>
+  );
+}
